+++ previous-versions/commerce/src/codatcommerce/models/operations/get_company_info.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import companyinfo as shared_companyinfo
-from ..shared import errormessage as shared_errormessage
+from ...models.shared import companyinfo as shared_companyinfo
 from typing import Optional
 
 
@@ -26,8 +25,6 @@ class GetCompanyInfoResponse:
     r"""HTTP response status code for this operation"""
     company_info: Optional[shared_companyinfo.CompanyInfo] = dataclasses.field(default=None)
     r"""OK"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""