+++ previous-versions/commerce/src/codatcommerce/models/operations/list_disputes.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import disputes as shared_disputes
-from ..shared import errormessage as shared_errormessage
+from ...models.shared import disputes as shared_disputes
 from typing import Optional
 
 
@@ -34,8 +33,6 @@ class ListDisputesResponse:
     r"""HTTP response status code for this operation"""
     disputes: Optional[shared_disputes.Disputes] = dataclasses.field(default=None)
     r"""OK"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your `query` parameter was not correctly formed"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""