+++ previous-versions/commerce/src/codatcommerce/models/operations/list_payments.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import payments as shared_payments
+from ...models.shared import payments as shared_payments
 from typing import Optional
 
 
@@ -32,8 +31,6 @@ class ListPaymentsResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your `query` parameter was not correctly formed"""
     payments: Optional[shared_payments.Payments] = dataclasses.field(default=None)
     r"""OK"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)