+++ previous-versions/commerce/src/codatcommerce/models/operations/list_tax_components.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import taxcomponents as shared_taxcomponents
+from ...models.shared import taxcomponents as shared_taxcomponents
 from typing import Optional
 
 
@@ -32,8 +31,6 @@ class ListTaxComponentsResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your `query` parameter was not correctly formed"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""
     tax_components: Optional[shared_taxcomponents.TaxComponents] = dataclasses.field(default=None)