+++ previous-versions/commerce/src/codatcommerce/models/shared/productinventory.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 import dataclasses
-from ..shared import productinventorylocation as shared_productinventorylocation
+from .productinventorylocation import ProductInventoryLocation
 from codatcommerce import utils
 from dataclasses_json import Undefined, dataclass_json
 from decimal import Decimal
@@ -13,7 +13,7 @@ from typing import List, Optional
 @dataclasses.dataclass
 class ProductInventory:
     r"""Information about the total inventory as well as the locations inventory is in."""
-    locations: Optional[List[shared_productinventorylocation.ProductInventoryLocation]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('locations'), 'exclude': lambda f: f is None }})
+    locations: Optional[List[ProductInventoryLocation]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('locations'), 'exclude': lambda f: f is None }})
     total_quantity: Optional[Decimal] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('totalQuantity'), 'encoder': utils.decimalencoder(True, False), 'decoder': utils.decimaldecoder }})
     r"""The total quantity of stock remaining across locations."""