+++ previous-versions/commerce/src/codatcommerce/sdk.py
@@ -13,7 +13,7 @@ from .tax_components import TaxComponents
 from .transactions import Transactions
 from codatcommerce import utils
 from codatcommerce.models import shared
-from typing import Dict
+from typing import Callable, Dict, Union
 
 class CodatCommerce:
     r"""Commerce API: Codat's standardized API for accessing commerce data
@@ -25,11 +25,11 @@ class CodatCommerce:
 
     [See our OpenAPI spec](https://github.com/codatio/oas)
     """
-    company_info: CompanyInfo
-    r"""Retrieve standardized data from linked commerce platforms."""
     customers: Customers
     r"""Retrieve standardized data from linked commerce platforms."""
     disputes: Disputes
+    r"""Retrieve standardized data from linked commerce platforms."""
+    company_info: CompanyInfo
     r"""Retrieve standardized data from linked commerce platforms."""
     locations: Locations
     r"""Retrieve standardized data from linked commerce platforms."""
@@ -47,7 +47,7 @@ class CodatCommerce:
     sdk_configuration: SDKConfiguration
 
     def __init__(self,
-                 auth_header: str,
+                 auth_header: Union[str,Callable[[], str]],
                  server_idx: int = None,
                  server_url: str = None,
                  url_params: Dict[str, str] = None,
@@ -57,7 +57,7 @@ class CodatCommerce:
         """Instantiates the SDK configuring it with the provided parameters.
         
         :param auth_header: The auth_header required for authentication
-        :type auth_header: str
+        :type auth_header: Union[str,Callable[[], str]]
         :param server_idx: The index of the server to use for all operations
         :type server_idx: int
         :param server_url: The server URL to use for all operations
@@ -72,22 +72,20 @@ class CodatCommerce:
         if client is None:
             client = requests_http.Session()
         
-        
-        security_client = utils.configure_security_client(client, shared.Security(auth_header = auth_header))
-        
+        security = shared.Security(auth_header = auth_header)
         
         if server_url is not None:
             if url_params is not None:
                 server_url = utils.template_url(server_url, url_params)
 
-        self.sdk_configuration = SDKConfiguration(client, security_client, server_url, server_idx, retry_config=retry_config)
+        self.sdk_configuration = SDKConfiguration(client, security, server_url, server_idx, retry_config=retry_config)
        
         self._init_sdks()
     
     def _init_sdks(self):
-        self.company_info = CompanyInfo(self.sdk_configuration)
         self.customers = Customers(self.sdk_configuration)
         self.disputes = Disputes(self.sdk_configuration)
+        self.company_info = CompanyInfo(self.sdk_configuration)
         self.locations = Locations(self.sdk_configuration)
         self.orders = Orders(self.sdk_configuration)
         self.payments = Payments(self.sdk_configuration)