+++ previous-versions/commerce/RELEASES.md
@@ -637,3 +637,13 @@ Based on:
 - [python v0.35.0] previous-versions/commerce
 ### Releases
 - [PyPI v0.35.0] https://pypi.org/project/codat-commerce/0.35.0 - previous-versions/commerce
+
+## 2023-11-13 01:29:13
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Commerce.yaml
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.36.0] previous-versions/commerce
+### Releases
+- [PyPI v0.36.0] https://pypi.org/project/codat-commerce/0.36.0 - previous-versions/commerce